pyproject.toml
[tool.poetry]
name = "pubtator"
version = "0.1.5"
description = ""
authors = ["James Morrill <james.morrill.6@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
requests = "^2.26.0"
numpy = "^1.21.0"
pre-commit = "^2.13.0"
biopython = "^1.79"
xmltodict = "^0.12.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 120

[tool.black]
line-length = 120
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''
